Refactor timers, divided code into separate modules
added rejected promise handling, added restart function for mods
